requirements.txt
Django==2.2.6
django-markdownx==2.0.28

